[FIXED SECURITY-53]
Administration page should be protected. Being an outside to this
plugin, it might not be the correct permission to check, but for now I'm
erring on the side of caution.

If you are maintaining this plugin, please feel free to introduce a new
permission for this, if Jenkins.ADMINISTER is too restrictive.